added some debug messages
