+++ go/cs/beaconing/mechanisms/pqa/mechanism.go
@@ -2,13 +2,10 @@ package pqa
 
 import (
 	"context"
-	"crypto/rand"
-	"math/big"
-	"time"
 
+	"github.com/scionproto/scion/go/cs/beacon"
 	"github.com/scionproto/scion/go/cs/beaconing"
 	"github.com/scionproto/scion/go/cs/ifstate"
-	"github.com/scionproto/scion/go/lib/ctrl/seg"
 	"github.com/scionproto/scion/go/lib/log"
 	"github.com/scionproto/scion/go/lib/serrors"
 )
@@ -24,52 +21,51 @@ type Mechanism struct {
 	Extender beaconing.Extender
 }
 
-func filterOverdue(ctx context.Context, tick beaconing.Tick, intfs []*ifstate.Interface) []*ifstate.Interface {
-	if tick.Passed() {
-		return intfs
+func (m Mechanism) ProvidePropagationBatch(ctx context.Context, tick beaconing.Tick) (beaconing.SendableBeaconsBatch, error) {
+	logger := log.FromCtx(ctx)
+	if !tick.Passed() {
+		return beaconing.SendableBeaconsBatch{}, nil
 	}
-	var stale []*ifstate.Interface
-	for _, intf := range intfs {
-		if tick.Overdue(intf.LastOriginate()) {
-			stale = append(stale, intf)
-		}
-	}
-	return stale
-}
-
-func (m *Mechanism) UpdateOriginationIntervals(ctx context.Context) {
-	for ifid := range m.Settings.Origination.Intervals {
-		if intf := m.AllInterfaces.Get(ifid); intf != nil {
-			if m.Tick.Overdue(intf.LastOriginate()) {
-				m.Settings.Origination.Intervals[ifid]++
+	res := make(beaconing.SendableBeaconsBatch, 0)
+	for _, srcIA := range m.getSourceIAs(ctx) {
+		for _, target := range m.getTargetsFromReceivedBeacons(ctx, srcIA) {
+			for _, neigh := range m.getNeighbouringASs(ctx) {
+				for _, intfSubgroup := range m.getIntfSubgroups(ctx, target, neigh) {
+					// logger.Debug("finding batch for", "src ia", srcIA, "target", target, "neigh", neigh, "intf subgroup", intfSubgroup)
+					bcns, err := m.getPropagationBatch(ctx, target, intfSubgroup, neigh, srcIA)
+					// logger.Debug("found batch:", "bcns", len(bcns))
+					if err != nil {
+						return nil, serrors.WrapStr("getting beacon batch", err)
+					}
+					for _, bcn := range bcns {
+						intf := m.AllInterfaces.Get(bcn.EgIfId)
+						res.AppendBeacon(intf, bcn)
+					}
+				}
 			}
 		}
 	}
+	logger.Debug("returning batch of size", "size", len(res))
+	return res, nil
 }
 
-func (m *Mechanism) getOptimizationTargetsForInterface(ctx context.Context, ifid uint16, interval uint) []Target {
-	logger := log.FromCtx(ctx)
-	origSettings := m.Settings.Origination
-	if order, ok := origSettings.Orders[ifid]; ok {
-		interval = interval % uint(len(order))
-		return origSettings.Orders[ifid][interval]
-	} else {
-		logger.Info("no target defined for interface", "ifid", ifid, "orders", origSettings.Orders)
-		return []Target{}
+func (m Mechanism) ProvideOriginationBatch(ctx context.Context, tick beaconing.Tick) (beaconing.SendableBeaconsBatch, error) {
+	m.Tick = tick
+	m.UpdateOriginationIntervals(ctx)
+	res := make(beaconing.SendableBeaconsBatch, 0)
+	intfsNeedingBeacons := filterOverdue(ctx, m.Tick, m.OriginationInterfaces())
+	for _, intf := range intfsNeedingBeacons {
+		intfId := intf.TopoInfo().ID
+		interval := m.Settings.Origination.Intervals[intfId]
+		targets := m.getOptimizationTargetsForInterface(ctx, intfId, interval)
+		res[intf] = make([]beacon.Beacon, 0)
+		for _, target := range targets {
+			bcn, err := m.CreateOriginBeaconForTarget(ctx, intfId, target)
+			if err != nil {
+				return nil, err
+			}
+			res[intf] = append(res[intf], bcn)
+		}
 	}
-
-}
-
-func (m *Mechanism) createSegment(ctx context.Context, timestamp time.Time, egIntfId uint16) (*seg.PathSegment, error) {
-	segID, err := rand.Int(rand.Reader, big.NewInt(1<<16))
-	if err != nil {
-		return nil, err
-	}
-
-	bseg, err := seg.CreateSegment(timestamp, uint16(segID.Uint64()))
-	if err != nil {
-		return nil, serrors.WrapStr("creating segment", err)
-	}
-
-	return bseg, nil
+	return res, nil
 }
