+++ go/cs/beaconing/mechanisms/pqa/propagation.go
@@ -5,41 +5,12 @@ import (
 	"sort"
 
 	"github.com/scionproto/scion/go/cs/beacon"
-	"github.com/scionproto/scion/go/cs/beaconing"
 	"github.com/scionproto/scion/go/cs/ifstate"
 	"github.com/scionproto/scion/go/lib/addr"
 	pqa_extension "github.com/scionproto/scion/go/lib/ctrl/seg/extensions/pqabeaconing"
 	"github.com/scionproto/scion/go/lib/log"
 	"github.com/scionproto/scion/go/lib/serrors"
 )
-
-func (m Mechanism) ProvidePropagationBatch(ctx context.Context, tick beaconing.Tick) (beaconing.SendableBeaconsBatch, error) {
-	logger := log.FromCtx(ctx)
-	if !tick.Passed() {
-		return beaconing.SendableBeaconsBatch{}, nil
-	}
-	res := make(beaconing.SendableBeaconsBatch, 0)
-	for _, srcIA := range m.getSourceIAs(ctx) {
-		for _, target := range m.getTargetsFromReceivesBeacons(ctx, srcIA) {
-			for _, neigh := range m.getNeighbouringASs(ctx) {
-				for _, intfSubgroup := range m.getIntfSubgroups(ctx, target, neigh) {
-					// logger.Debug("finding batch for", "src ia", srcIA, "target", target, "neigh", neigh, "intf subgroup", intfSubgroup)
-					bcns, err := m.getPropagationBatch(ctx, target, intfSubgroup, neigh, srcIA)
-					// logger.Debug("found batch:", "bcns", len(bcns))
-					if err != nil {
-						return nil, serrors.WrapStr("getting beacon batch", err)
-					}
-					for _, bcn := range bcns {
-						intf := m.AllInterfaces.Get(bcn.EgIfId)
-						res.AppendBeacon(intf, bcn)
-					}
-				}
-			}
-		}
-	}
-	logger.Debug("returning batch of size", "size", len(res))
-	return res, nil
-}
 
 // Returns the propagation batch as proposed by Ali et.al. for PQA Beaconing
 func (m Mechanism) getPropagationBatch(ctx context.Context, target Target, egIntfG []*ifstate.Interface, neigh addr.IA, src addr.IA) ([]beacon.Beacon, error) {
@@ -114,7 +85,7 @@ func (m Mechanism) getSourceIAs(ctx context.Context) []addr.IA {
 	return addrs
 }
 
-func (m Mechanism) getTargetsFromReceivesBeacons(ctx context.Context, srcIA addr.IA) []Target {
+func (m Mechanism) getTargetsFromReceivedBeacons(ctx context.Context, srcIA addr.IA) []Target {
 	targets, err := m.GetActiveTargets(ctx, srcIA)
 	if err != nil {
 		log.FromCtx(ctx).Error("error getting active targets", "err", err)
